% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABA.R
\docType{data}
\name{ABA}
\alias{ABA}
\title{Data from a reversal design.}
\format{
A \code{data.frame} of 12 rows and 3 columns.
\describe{
 \item{Session}{Numeric values giving number of either, day or session.}
 \item{Measure}{Numeric values giving the magnitude of dependent variable for each day or session.}
 \item{Phase}{String values giving the label for each phase.}
   
}
}
\source{
Data obtained by computer simulation.
}
\usage{
ABA
}
\description{
Data set exemplifying the organization of data for behavior registered through 12 sessions that included a baseline, an intervention and a reversion to baseline.
}
\keyword{datasets}
