% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{Data from a self-regulated study session on a computerized task}
\format{
A \code{data.frame} of 35 rows and 3 columns.
\describe{
 \item{Serie}{String values corresponding to several categories registered through time of observation.}
 \item{Inicio}{Numeric values indicating the start time of each category.}
 \item{Final}{Numeric values indicating the end time of each category.}
   
}
}
\source{
Data obtained during a test of the programmed task.
}
\usage{
categories
}
\description{
Data set exemplifying the organization of data for several categories of interactions with material on logic gates in a computerized task.
}
\keyword{datasets}
