% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumRecord.R
\name{cumRecord}
\alias{cumRecord}
\title{Cumulative record}
\usage{
cumRecord(
  time,
  events,
  t_res = 1,
  code = c(1, 2),
  ymax = 100,
  xlab = "Time in seconds",
  ylab = "Cumulative responses"
)
}
\arguments{
\item{time}{vector that contains moments in which some type of event occurred.}

\item{events}{vector that contains the sequence of numeric values corresponding to registered events.}

\item{t_res}{number to indicate the resolution of time. By defult \code{t_res=1}, for a resolution of time in seconds.}

\item{code}{vector with values used to identify the correspondence between numbers and the type of events they refer to, for example: (1) responses and (2) reinforcers. By default \code{code=c(1,2)}.}

\item{ymax}{value to establish the maximum in Y axis.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis}
}
\value{
Returns a graphic that shows cumulative responses over time of observation.
}
\description{
Plot cumulative responses through consecutive seconds (or any other unit)  of a session.
}
\examples{
cumRecord(operants$time,operants$events,t_res=.2,code=c(1,7),ymax=100,xlab="Time in seconds",ylab="Cumulative responses")
 
}
\author{
Jorge A. Ruiz <ruizvja@yahoo.com>, Cristina Guerrero <gusc1001@gmail.com>
}
