% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl.R
\docType{data}
\name{mbl}
\alias{mbl}
\title{Data from a multiple baseline design}
\format{
A \code{data.frame} of 60 rows and 4 columns.
\describe{
 \item{Session}{Numeric values giving number of either, day or session.}
 \item{Measure}{Numeric values giving the magnitude of dependent variable for each day or session.}
 \item{Phase}{String values giving the label for each phase.}
 \item{Case}{String values giving the corresponding label of each case.}
 
}
}
\source{
Randomly generated data
}
\usage{
mbl
}
\description{
A fictitious data set exemplifying the organization of data for a multiple baseline design with three cases in two phases, lasting 60 consecutive sessions.
}
\keyword{datasets}
