% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multielement.R
\docType{data}
\name{multielement}
\alias{multielement}
\title{Data from a multielement design}
\format{
A \code{data.frame} of 25 rows and 4 columns.
\describe{
 \item{Session}{Numeric values giving number of either, day or session.}
 \item{Measure}{Numeric values giving the magnitude of dependent variable for each day or session.}
 \item{Phase}{String values giving the label for each phase.}
 \item{Treatment}{String values giving the corresponding label of data under each treatment.}
 
}
}
\source{
Randomly generated data
}
\usage{
multielement
}
\description{
A fictitious data set exemplifying the organization of data for a multielement design with three phases, linebase, intervention 1 and intervention 2, which lasted 5, 10 and 10 sessions each one.
}
\keyword{datasets}
