% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categories.R
\name{plot_categories}
\alias{plot_categories}
\title{Behavior categories in continuous time}
\usage{
plot_categories(
  categories,
  start,
  end,
  lev = NULL,
  xlab = "Time of observation",
  ylab = NULL,
  title = "title",
  col = I("black")
)
}
\arguments{
\item{categories}{vector that contains the sequence of numeric or character values corresponding to registered categories.}

\item{start}{vector with values to indicate time of beginning of each category.}

\item{end}{vector with values to indicate the time in which each category ends.}

\item{lev}{vector with values used to indicate an order of to indicate the order of appearance of the categories on the Y axis. By default, this argument is NULL, so the categories are placed in ascending order.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{title}{text to show as title of plot.}

\item{col}{value or vector to indicate color for categories in the plot. By default \code{col=I("black")}.}
}
\value{
Returns a graphic that shows categories of behavior over time of observation.
}
\description{
Plot the occurrence of behavior categories through consecutive time (in any unit) of a period of observation.
}
\examples{
plot_categories(categories$Serie,categories$Inicio,categories$Final) 

plot_categories(categories$Serie,categories$Inicio,categories$Final,lev=c("O","e","E","h","L","s"),col=c(categories$Serie,categories$Serie))
 
}
\author{
Jorge A. Ruiz <ruizvja@yahoo.com>
}
