% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dir.R
\name{plot_dir}
\alias{plot_dir}
\title{Plot routes per cycle}
\usage{
plot_dir(
  x,
  y,
  n.cycles,
  len.cycles,
  xlim,
  ylim,
  xlab = "X",
  ylab = "Y",
  roi = c(xlim[2]/2, ylim[2]/2),
  r = ylim[2]/10
)
}
\arguments{
\item{x}{vector that contains coordinates in x axis.}

\item{y}{vector that contains coordinates in y axis.}

\item{n.cycles}{value used to indicate the number of cycles.}

\item{len.cycles}{value used to indicate the length of each cycle.}

\item{xlim}{vector with min and max values for x axis.}

\item{ylim}{vector with min and max values for y axis.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{roi}{vector with values to indicate coordinates of a region of interest.}

\item{r}{value to indicate the radius of ROI.}
}
\value{
Returns a graphic that shows categories of behavior over time of observation.
}
\description{
Plot the route correspondiente a cada uno de los ciclos en los que se divide el tiempo de registro.
}
\examples{
plot_categories(categories$Serie,categories$Inicio,categories$Final) 

plot_categories(categories$Serie,categories$Inicio,categories$Final,lev=c("O","e","E","h","L","s"),col=c(categories$Serie,categories$Serie))
 
}
\author{
Jorge A. Ruiz <ruizvja@yahoo.com>
}
