% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dirJPG.R
\name{plot_dirJPG}
\alias{plot_dirJPG}
\title{Plot routes per cycle (or trial)}
\usage{
plot_dirJPG(
  x,
  y,
  cycle,
  select = 0,
  xlim,
  ylim,
  xlab = "X",
  ylab = "Y",
  imgs = NULL,
  roi = c(xlim[2]/2, ylim[2]/2),
  r = ylim[2]/10
)
}
\arguments{
\item{x}{vector that contains coordinates in x axis, frame by frame.}

\item{y}{vector that contains coordinates in y axis, frame by frame.}

\item{cycle}{vector with values indicating the corresponding number of cycle or trial for each frame.}

\item{select}{value or vector with values indicating the number of cycle or trial to be plotted. By default, \code{select=0} produce a plot for each temporal segment.}

\item{xlim}{vector with min and max values for x axis.}

\item{ylim}{vector with min and max values for y axis.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{roi}{vector with a pair of values to indicate coordinates of a region of interest (assuming a circular shape).}

\item{r}{value to indicate the radius of \code{roi}.}
}
\value{
Returns a graphic that shows routes of movement through the experimental arena, by cycle or by trial.
}
\description{
Plot the routes corresponding to the several cycles or trials within the session time.
}
\examples{
library(interp)  #install interp package if needed
x11()
par(pty="s")
par(mfrow=c(3,3))
plot_dir(xy_test1$x,xy_test1$y,xy_test1$cycle,select=c(1,2,3,43,44,45,70,71,72),xlim=c(0,92),ylim=c(0,92),roi=c(0,46))

 
}
\author{
omitted for blind review
}
