% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MBL.R
\name{plot_MBL}
\alias{plot_MBL}
\title{Plot data from a multiple baseline design}
\usage{
plot_MBL(
  time,
  dimension,
  case,
  phase,
  symbol = 11,
  type = "o",
  xlab = "x",
  ylab = "Y-axis",
  xdiv = 0,
  ylim = NULL
)
}
\arguments{
\item{time}{vector that contains the number of day or session.}

\item{dimension}{vector that contains the values to be plotted for each day or session.}

\item{case}{vector with labels indicating each case.}

\item{phase}{vector with the labels of corresponding phase for each day or session.}

\item{symbol}{symbol to use for plotting data. By default \code{pch=11},but it can be changed by the user.}

\item{type}{character that indicates to plot both, points and line. By default, \code{type="o"} (overplotted), but it can be changed for \code{type="b"} (with breaks between each point and the line).}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{xdiv}{value to indicate the number of intervals into which the x-axis is divided. By default, value 0 results in dividing the axis into n-1 intervals, where n is the total number of sessions.}

\item{ylim}{numeric value, giving the y-axis range. By default, \code{ylim=NULL} sets the limit in the next ten with respect to the maximum in \code{dimension} variable.}
}
\value{
Returns a graphic that shows the registered value for each day, indicating the current phase, as well as the change from one phase to another for each case.
}
\description{
Plot data through consecutive days or sessions, indicating the label of each phase, the case to whom each registered data belongs, and when changes of conditions occurred for each case.
}
\examples{
plot_MBL(mbl$session,mbl$measure,mbl$case,mbl$phase,xdiv=5,ylim=NULL) 


}
\author{
omitted for blind review
}
