% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ME.R
\name{plot_ME}
\alias{plot_ME}
\title{Plot data from a multielement design}
\usage{
plot_ME(
  time,
  dimension,
  treatment,
  phase,
  symbol = c(0, 15, 19),
  type = "o",
  xlab = "x",
  ylab = "y",
  legend = "Subject",
  xdiv = 0,
  ylim = NULL
)
}
\arguments{
\item{time}{vector that contains the number of day or session.}

\item{dimension}{vector that contains the values to be plotted for each day or session.}

\item{treatment}{vector with labels indicating the corresponding treatment for each day.}

\item{phase}{vector with the labels of corresponding phase for each day or session.}

\item{symbol}{symbol to use for plotting data. By default \code{pch=18},but it can be changed by the user.}

\item{type}{character that indicates to plot both, points and line. By default, \code{type="o"} (overplotted), but it can be changed for \code{type="b"} (with breaks between each point and the line).}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis}

\item{legend}{label to identify the subject.}

\item{xdiv}{value to indicate the number of intervals into which the x-axis is divided. By default, value 0 results in dividing the axis into n-1 intervals, where n is the total number of sessions.}

\item{ylim}{numeric value, giving the y-axis range. By default, \code{ylim=NULL} sets the limit in the next ten with respect to the maximum in \code{dimension} variable.}
}
\value{
Returns a graphic that shows the registered value for each day, indicating the current phase and treatment, as well as the change from one phase to another.
}
\description{
Plot data through consecutive days or sessions, indicating the label of each phase, the series for each alternating treatments, and when changes of conditions occurred.
}
\examples{
plot_ME(multielement$Session,multielement$Measure,multielement$Treatment,multielement$Phase,xdiv=5,ylim=NULL)


}
\author{
omitted for blind review
}
