% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDARBS.R
\docType{package}
\name{RDARBS}
\alias{RDARBS}
\alias{RDARBS-package}
\title{R package for data analysis and representation in behavior science}
\description{
Toolbox for easily analyze and visualize data from several types of research designs in behavior science
}
\author{
omitted for blind review
}
