% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CC.R
\name{plot_CC}
\alias{plot_CC}
\title{Plot data from a changing-criterion design}
\usage{
plot_CC(
  time,
  dimension,
  criteria,
  phase,
  symbol = 18,
  type = "o",
  xlab = "x",
  ylab = "y",
  legend = "Subject"
)
}
\arguments{
\item{time}{vector that contains the number of day or session.}

\item{dimension}{vector that contains the values to be plotted for each day or session.}

\item{criteria}{vector with the names of criteria for each day or session.}

\item{phase}{vector with the labels of corresponding phase for each day or session.}

\item{symbol}{symbol to use for plotting data. By default \code{pch=18},but it can be changed by the user.}

\item{type}{character that indicates to plot both, points and line. By default, \code{type="o"} (overplotted), but it can be changed for \code{type="b"} (with breaks between each point and the line).}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis}

\item{legend}{label to identify the subject.}
}
\value{
Returns a graphic that shows the registered value for each day, indicating the current criteria in each phase, as well as the change from one phase to another.
}
\description{
Plot data through consecutive days or sessions, indicating which was the current criteria in each phase and when each phase change occurred.
}
\examples{
plot_CC(changing$Session,changing$Measure,changing$Criteria,changing$Phase)
 
}
\author{
omitted for blind review
}
